# Backend/CIT-Forums/src/main/resources/static/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CIT Forums API",
    "description": "API documentation for CIT Forums platform",
    "version": "1.0.0",
    "contact": {
      "name": "CIT Forums Admin",
      "email": "admin@citforums.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/api/forums": {
      "get": {
        "summary": "Get all forums",
        "description": "Returns a paginated list of all forums",
        "tags": ["Forums"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (0-based)",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of forums",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedForumResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new forum",
        "description": "Creates a new forum with the given details",
        "tags": ["Forums"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForumRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Forum created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForumResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/forums/{id}": {
      "get": {
        "summary": "Get forum by ID",
        "description": "Returns a forum by its ID",
        "tags": ["Forums"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Forum ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Forum found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForumResponse"
                }
              }
            }
          },
          "404": {
            "description": "Forum not found"
          }
        }
      },
      "put": {
        "summary": "Update forum",
        "description": "Updates an existing forum",
        "tags": ["Forums"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Forum ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForumRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Forum updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForumResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Forum not found"
          }
        }
      },
      "delete": {
        "summary": "Delete forum",
        "description": "Deletes an existing forum",
        "tags": ["Forums"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Forum ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Forum deleted successfully"
          },
          "404": {
            "description": "Forum not found"
          }
        }
      }
    },
    "/api/threads": {
      "get": {
        "summary": "Get all threads",
        "description": "Returns a paginated list of all threads",
        "tags": ["Threads"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (0-based)",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of threads",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedThreadResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/threads/forum/{forumId}": {
      "get": {
        "summary": "Get threads by forum",
        "description": "Returns a paginated list of threads in a specific forum",
        "tags": ["Threads"],
        "parameters": [
          {
            "name": "forumId",
            "in": "path",
            "required": true,
            "description": "Forum ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (0-based)",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of threads",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedThreadResponse"
                }
              }
            }
          },
          "404": {
            "description": "Forum not found"
          }
        }
      },
      "post": {
        "summary": "Create thread in forum",
        "description": "Creates a new thread in a specific forum",
        "tags": ["Threads"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "forumId",
            "in": "path",
            "required": true,
            "description": "Forum ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Thread created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Forum not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "ForumRequest": {
        "type": "object",
        "required": ["title", "description", "category"],
        "properties": {
          "title": {
            "type": "string",
            "example": "General Discussion"
          },
          "description": {
            "type": "string",
            "example": "A place for general discussions"
          },
          "category": {
            "type": "string",
            "enum": ["ANNOUNCEMENTS", "EVENTS", "FREEDOM_WALL", "CONFESSION", "ACADEMIC", "GENERAL"],
            "example": "GENERAL"
          }
        }
      },
      "ForumResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "General Discussion"
          },
          "description": {
            "type": "string",
            "example": "A place for general discussions"
          },
          "category": {
            "type": "string",
            "example": "GENERAL"
          },
          "active": {
            "type": "boolean",
            "example": true
          },
          "threadCount": {
            "type": "integer",
            "example": 10
          },
          "createdBy": {
            "type": "string",
            "example": "admin@example.com"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PagedForumResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ForumResponse"
            }
          },
          "totalElements": {
            "type": "integer",
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "example": 10
          },
          "size": {
            "type": "integer",
            "example": 10
          },
          "number": {
            "type": "integer",
            "example": 0
          }
        }
      },
      "ThreadRequest": {
        "type": "object",
        "required": ["title", "content"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Welcome to the forum!"
          },
          "content": {
            "type": "string",
            "example": "This is the first post in our new forum."
          }
        }
      },
      "ThreadResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Welcome to the forum!"
          },
          "content": {
            "type": "string",
            "example": "This is the first post in our new forum."
          },
          "forumId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "commentCount": {
            "type": "integer",
            "example": 5
          },
          "createdBy": {
            "type": "string",
            "example": "user@example.com"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PagedThreadResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThreadResponse"
            }
          },
          "totalElements": {
            "type": "integer",
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "example": 10
          },
          "size": {
            "type": "integer",
            "example": 10
          },
          "number": {
            "type": "integer",
            "example": 0
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 